Add files via upload
This is the 1/32 micro stepping variant of the Christopher Stahl 7x7 firmware for the Bondtech BMG. Keeps the stock MK3 1/32 micro stepping for the extruder axis for those who want to use 830 e-steps on the extruder with a 1.8 degree motor or 1660 steps with a 0.9 degree motor. Changes made to firmware for better unload and load with the new Bondtech extruder.