Add files via upload
Here is the 1/32 micro stepping variant of the Bondtech extruder firmware. This has the same Christopher Stahl 7x7 bed leveling plus the added load and unload lengths to compensate for the Bondtech gears. The only difference is this is set for 1.8 degree motors being used on the Bondtech extruder. It will have 830 e-steps with a 1.8 degree motor OR if you'd like to use a 0.9 degree motor the e-steps need to be change to 1660.